requirement.txt
absl-py==0.13.0
aniso8601==9.0.1
APScheduler==3.7.0
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.6.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
conda==4.3.16
cycler==0.10.0
dataclasses==0.8
elasticsearch==7.13.2
firebase-admin==5.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gast==0.3.3
gensim==4.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
h5py==2.10.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
jamo==0.4.1
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
krwordrank==1.0.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mecab-python===0.996-ko-0.9.2
msgpack==1.0.2
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
Pillow==8.3.0
pkg-resources==0.0.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pymongo==3.11.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.4.1
six==1.16.0
smart-open==5.1.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.1
tweepy==3.10.0
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
wordcloud==1.8.1
wrapt==1.12.1
zipp==3.4.1
